refactor: move genkit flows to dedicated directory
